requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.1
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
coverage==6.2
Django==3.2.9
djangorestframework==3.12.4
frozenlist==1.2.0
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
multidict==5.2.0
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.2
py==1.11.0
pycrunch-engine==1.3.1
pyparsing==3.0.6
pytest==6.2.5
python-engineio==3.14.2
python-socketio==4.6.1
pytz==2021.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
watchgod==0.7
yarl==1.7.2
